Add files via upload
Displayed below are two-by-side interactive maps showcasing state data across the United States. Users can zoom in and pan around both maps. On the left, a bubble map leverages Plotly’s scattergeo map library to visualize the pollution data for U.S. states. Each of the four pollutants is depicted by colored circles, with the size of the circle corresponding to the pollutant value. Users can choose which pollutants they’d like to display by clicking on the legend. On the map to the right, the median income for U.S. states is displayed through a choropleth map. In both maps, hovering over any state triggers a tooltip displaying the respective value (such as pollutant air quality index or median income), the state name, and the year. At the top left corner, users can click on a drop-down menu to select a specific year ranging from 2000 to 2018, and the maps will repopulate with data from the selected year.